+++ analysis/src/Hover.ml
@@ -26,7 +26,10 @@ let showModuleTopLevel ~docstring ~isType ~name (topLevel : Module.item list) =
   let doc =
     match docstring with
     | [] -> ""
-    | _ :: _ -> "\n" ^ (docstring |> String.concat "\n") ^ "\n"
+    | _ :: _ ->
+      "\n"
+      ^ (docstring |> String.concat "\n")
+      ^ Markdown.divider ^ Markdown.spacing
   in
   Some (doc ^ full)