+++ CHANGELOG.md
@@ -18,6 +18,7 @@
 
 #### :rocket: New Feature
 
+- Add markdown divider between module doc and module type in hover information. https://github.com/rescript-lang/rescript/pull/7775
 - Show docstrings before type expansions on hover. https://github.com/rescript-lang/rescript/pull/7774
 
 #### :bug: Bug fix